refactor: 도메인별 DTO 패키지 구조 정리
