+++ sensor/src/main/java/com/iroom/sensor/dto/HeavyEquipment/EquipmentLocationUpdateResponse.java
@@ -1,4 +1,4 @@
-package com.iroom.sensor.dto;
+package com.iroom.sensor.dto.HeavyEquipment;
 
 public record EquipmentLocationUpdateResponse(
         Long id,
