+++ sensor/src/main/java/com/iroom/sensor/dto/HeavyEquipment/EquipmentRegisterResponse.java
@@ -1,4 +1,4 @@
-package com.iroom.sensor.dto;
+package com.iroom.sensor.dto.HeavyEquipment;
 
 import com.iroom.sensor.entity.HeavyEquipment;
